requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.9.0
psycopg2-binary==2.9.1
PyJWT==2.3.0
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
django-heroku==0.3.1